Use default phabricator api key to query public revisions
After a long talk with our security folks, we came to a conclusion that
having every user of Lando (1000+) enter their Phabricator API keys would
be a possible security risk, even if stored in a secure cookie.

This patch allows Lando to use 1 key that it owns itself to query for
public revisions on behalf of users. If the user does provide an api key,
that key will be used instead, which allows users who have access to secure
revisions to only input their API key when needed.