Fix Dockerfile-dev so it creates the lando-api-dev script
pip won't generate scripts listed in setup.py's entry_points if the
module is already available in the PYTHONPATH.  Scripts in the dev
environment, like lando-api-dev, are not being generated as a result.
Change when we set the PYTHONPATH environment variable in the dev
environment Dockerfile so pip will generate the scripts once
again.