Refactor python requirements and dev Docker image
As our application grows, we are encountering cases where not every
python reqirement is needed in each environment.

This commit is meant to orgainize our requirments in a simple folder
that makes it easy to see what is needed in each environment.

In addition, before our development image was using a requirement for
production (uwsgi) which made building the container take a very long
time since it has to compile a bunch of dependencies. Moving the uwsgi
requirment into it's own production requirements file allows the dev
Dockerfile to be simplified again.