Refactor app to use the /differential.querydiffs Phabricator API endpoint
Before Lando API used a special Phabricator API endpoint (/phid.query)
to get basic metadata about a diff - like its id. At the time this was
all we needed to support what we thought would be our workflow. Because
of this, the simple metadata object that Phabricator returned was known
as a "Diff" in the Lando API codebase, even though it was merely a
subset of the data.

Our requirements changed and we now need the full diff information
provided at Phabricator's /differential.querydiffs API endpoint. It
would not make sense two have 2 different types of "Diffs" in the
codebase, so this patch refactors all uses of "Diffs" to mean the
full diff information.

This patch refactors the PhabricatorClient and all usages of the old
diff methods. It also refactors the PhabricatorFactory and a lot of
the test suite since they are so tied to the PhabricatorClient.