Saving patch in S3
Currently patch's content was sent as a parameter of a POST request to
Transplant's `autoland` endpoint. Since patch can and often is a large
block of text, it has been decided to save it in S3 bucket.

The name of the bucket is to be saved in environent variable.
One might set the access and secret keys in AWS_ACCESS_KEY and
AWS_SECRET_KEY environment variables.