Refuse to land an inactive revision unless forced.
Summary:
Lando API will fail with a 409 error if diff used to land revision is not the active one. Client will then provide an information about the issue and will give an option to force landing the inactive diff. If user decides to do so client will request a new landing with a `force_override_of_diff_id` parameter equal to the active id. API will proceed with the landing.

For the record an active diff id is saved in database.

Additional fixes:
    - Fixed creation of phid response in `tests.factories`
    - Renamed the `land` method to `post` in `api.landings`
    - Removed dependency on `json` in `api.landings`
    - Added a msg to the RevisionNotFoundException
    - Removed space at the end of the line in swagger

Test Plan:
invoke test,
invoke upgrade before testing in UI

Reviewers: imadueme, dkl, smacleod

Reviewed By: imadueme, smacleod

Bug #: 1396602

Differential Revision: https://phabricator.services.mozilla.com/D38