Bug 1399219 - Add commit_message_preview field to GET /revisions endpoint
Summary:
Adds a new field which holds the default commit message that will be used when the Revision is landed. The default message is included so that clients are able to show the user a preview of the commit message that will be used when landing the Revision onto the target repo. This will be the default message and from there clients are able to allow the user to edit the message before landing. The default commit message takes on the format of:

`<Bug #> - <Commit Message Title> - <r=reviewer1,...>`

This Revision only adds the basic support for getting this field. [[ https://bugzilla.mozilla.org/show_bug.cgi?id=1400306 | Bug 1400306 ]] tracks adding the mozautomation.commitparser library (https://hg.mozilla.org/hgcustom/version-control-tools/file/tip/pylib/mozautomation/mozautomation/commitparser.py) to make the validation of commit messages stronger and more accurate.

Test Plan:
  - Run invoke test to ensure all unit tests pass.
  - Run the API server and try getting a live revision. It should include the commit_message_preview field in the response.

Reviewers: zalun

Reviewed By: zalun

Bug #: 1399219

Differential Revision: https://phabricator.services.mozilla.com/D57