auth0: add require_access_token decorator.
Summary:
This change adds a require_access_token decorator meant to decorate
connexion endpoint handler functions. This decorator will require
the Authorization header be set properly with an auth0 jwt Bearer
token.

To properly verify the signature used for the access_tokens two
new environment variables are required, OIDC_IDENTIFIER and
OIDC_DOMAIN, set to the auth0 custom api identifier name and the
auth0 domain for the custom api respectively.

Missing from this change is an actual endpoint protected by the
decorator which will come in a future change. Also missing is the
ability to use the verified access token to contact auth0 and
request extra user information - this will come in a future
change as well.

Test Plan:
New tests for the provided decorator are included and `invoke test`
passes.

Reviewers: davidwalsh, imadueme

Reviewed By: davidwalsh, imadueme

Subscribers: imadueme

Bug #: 1411373

Differential Revision: https://phabricator.services.mozilla.com/D155