auth0: add require_auth0_userinfo decorator.
Summary:
This change introduces the require_auth0_userinfo decorator which
builds on the require_access_token decorator. It should be used
for endpoints which require full user profile information from
auth0 rather than just a verified jwt.

The require_auth0_userinfo decorator will first verify the access_token
using require_access_token and then use it to request userinfo from
auth0. It is possible auth0 rejects a valid access_token if it was
not requested with the proper scope / audience (openid scope is
required to generate a token with the https://<auth0_domain>/userinfo
audience.

According to the documentation auth0 will always return custom claim
values from when the access_token was generated, not when the
/userinfo endpoint is called. This means subsequent calls to /userinfo
will always return the same custom claim values even if they have
changed. Since the claims we care about are custom claims the first
call to /userinfo for a given token could be cached, but that is
left to a future change.

Test Plan:
New tests for the provided decorator are included and `invoke test`
passes.

Reviewers: davidwalsh, imadueme

Reviewed By: davidwalsh, imadueme

Bug #: 1411373

Differential Revision: https://phabricator.services.mozilla.com/D160