api: move phab api key to request header and check using a decorator
Summary:
The api sometimes requires a Phabricator Conduit api key to perfom
some operations. Before this commit it was accepted via a query
parameter which could cause some unintended exposure of the api key.
This commit moves it to a more secure and standard X-PhabricatorAPI-Key
header using the security definitions feature of the OpenAPI Spec.

This commit also adds a decorator: require_phabricator_api_key
which makes it easy for functions to easily obtain a PhabricatorClient
with the correct permission level and the decorator handles responding
to the user in case of errors concerning the api key.

https://github.com/mozilla-conduit/packrat by smacleod was the basis
of much of this work.

Test Plan:
Run invoke test. If you have your local lando-api setup to
talk to a Phabricator instance, you may also test it via the Swagger UI
by trying to access a secure Revision without the api key set and with
the api key set.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1409443

Differential Revision: https://phabricator.services.mozilla.com/D189