Landing model: Add timestamps
Summary:
We're not storing the time when the Landing object was saved. This patch is
adding the time of creation and the update time.

`created_at` and `updated_at` DateTime attributes are added to Landing mdoel.
They're set in `__init__` and `save` methods to the current UTC time.
freezegun is used for testing.

Reviewers: imadueme, smacleod

Reviewed By: imadueme, smacleod

Bug #: 1413917

Differential Revision: https://phabricator.services.mozilla.com/D184