landing model: add requester_email and tree as fields
Summary:
The email of the person who requested the revision and the tree
which the revision will land to are necessary for clients to know when
they retreive landing information.

Test Plan:
Run invoke upgrade to ensure the migrations run properly.
You can try a fresh db by starting postgres and dropping the database.
Run invoke test. You can also startup the api, connect it to a phabricator
instance, disable the check in api/landings.py for an auth0 user, and
manually test using the swagger ui.

Reviewers: zalun

Reviewed By: zalun

Differential Revision: https://phabricator.services.mozilla.com/D240