landing: send proper `rev` to transplant.
Summary:
The transplant server uses the `rev` argument as the primary key for
landing requests. This means it needs to be unique for each landing
request, but equal where multiple requests should cause a failure due
to duplication prevention. The obvious choice here is the revision ID
itself, as we don't want to allow a request to land a revision which
is already in progress.

In the future this choice of a single revision identifier for landing
multiple patches will be problematic; Since the dependency
relationship for revisions is mutable just using the "tip" revision
will not work as it did for immutable commits. Multiple patch landing
introduces a lot of extra complexity so I'm going to leave fixing
this issue to future work.

MozReview fills the 'rev' field with a commit sha in hex format, so
it's theoritically possible that a string of the form "D<integer>"
could collide, if the integer is greater than 10 billion (1*10^10) and
short commit identifiers were being used. This will never happen in
practice though as it would require 10 revisions be created per second
for the next 30 years or so.

We choose to prefix the revision integer id with a "D" as this is the
standard format for display and is much less ambiguous.

Test Plan: `$ invoke test` passes.

Reviewers: imadueme, glob

Reviewed By: imadueme, glob

Differential Revision: https://phabricator.services.mozilla.com/D315