Landing: Storing revision id as integer.
Summary:
Our system is inconsistent about the format of the `revision_id`. Sometimes
it is used as an integer and another as a string. This patch is changing
the behaviour to use integer inside the application. On the "outside"
`revision_id` is presented as a string in the format of `'D{number}'`.
Value is translated from string to integer on the input and translated
back to the string on the output.

Reviewers: imadueme, dylan

Reviewed By: imadueme, dylan

Bug #: 1418400

Differential Revision: https://phabricator.services.mozilla.com/D255