Patch: Check if diff is linked to the right revision
Summary:
The user is choosing a diff and revision to be landed. We are not checking if
the diff is actually related to the requested revision. This patch fixes
this issue.

New method `Patch.check_integrity` is added. It is called when Patch is
initialized. It checks if `revision['diffs']` list contains the required `diff_id`.

Other fixes.
In `tests/factories.py` `str(diff_id)` is used as a key. Using integer was
raising a 404. `differential.querydiffs` is returning a JSON with a dict and
we check the key to find the diff. We haven't seen this issue before because
we were using only `differential.getrawdiff` API.

Nits fixed:
api/landings.py style fix
tests/test_landings.py removed `import pytest`

Refactoring:
tests/factories - revision is using named kwargs

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1416804

Differential Revision: https://phabricator.services.mozilla.com/D227