auth/landings: Add A0User email property and require it for landing
Summary:
Provides the email property on the Auth0 user class which internally
looks for the user's email in the 'email' field of the userinfo and
checks the 'email_verified' field to ensure it is valid.

Having a valid email, is now a requirement for requesting a landing.
Otherwise the Transplant service would not have an email to use as the
push user which could lead to pushes with no one to trace back to.

Test Plan: Run invoke test.

Reviewers: smacleod

Reviewed By: smacleod

Differential Revision: https://phabricator.services.mozilla.com/D317