landing: send proper 'destination' to transplant.
Summary:
This change updates the data lando sends to transplant when requesting
a landing to remove uneeded fields and send a proper 'destination'
field. As the included code comment indicates, we can get away
with just using 'upstream' as the destination since we don't currently
support landing to try.

'push_bookmark' is also no longer sent, as it is not used for
mozilla-central and adding proper support for the repos that need it
would increase this changes complexity quite a bit. We can circle back
around and add proper push_bookmark support once it will be used.

Test Plan: `$ invoke test` passes

Reviewers: imadueme, glob

Reviewed By: imadueme, glob

Differential Revision: https://phabricator.services.mozilla.com/D324