auth/transplant: Add localdev options to mock restricted actions
Summary:
Auth0 and Transplant are both external services lando-api depends on
to land things for users. In our local development environment, there
was no way to simulate landing a change without manually going into the
code to remove some of the hard checks/requests made to Auth0 and Transplant.

This commit aims to selectively mock some of these behaviors in localdev
by introducing a set of localdev mock flags which hopefully will grow
to a full suit of flags for a highly configurable localdev environment.

The two new flags added are documented in the docker-compose.yml,
one to mock Auth0 users and another to mock the submission request
to Transplant.

A new 'landoapi.mocks' module is added to support these changes.

Test Plan: See the attached video.

Reviewers: smacleod

Reviewed By: smacleod

Differential Revision: https://phabricator.services.mozilla.com/D341