landings api: remove 'status' filter from the list landings endpoint
Summary:
The status filter allowed us to view all landings with a particular
status. This was never being used and in hindsight if it were used
it would be problematic since it didn't limit the amount of landings
that could be returned.

In preperation for securing the get landing list endpoint this filter
needed to be removed since we would otherwise need to verify that the
client is able to see every single revision associated with each landing.

Test Plan:
run invoke test and feel free to fire up the api and try
it yourself via the swagger ui helper.

Reviewers: dylan

Reviewed By: dylan

Differential Revision: https://phabricator.services.mozilla.com/D363