landings: secure the /landings and /landings/{id} endpoints
Summary:
This patch makes sure that clients are able to view the revision
associated with a landing before returning landing information
for that revision.

Tests were updated to reflect this change and new tests were added
for the new features.

Test Plan: invoke test and try it yourself via the swagger ui

Reviewers: mars

Reviewed By: mars

Differential Revision: https://phabricator.services.mozilla.com/D380