revisions: add 'latest_diff_id' field to /revisions/{id} endpoint
Summary:
In a prior patch the 'diff_id' field was added to the /revisions/{id}
endpoint which allowed clients to request revision information in the
context of a specific diff (perhaps an older diff). When this happened
clients lost the ability to know what the latest diff was (since before
the latest diff was always returned).

This patch adds the latest diff id field so that clients can check
if the user is working with the latest diff and take any actions based
upon that.

Test Plan:
invoke test and try it out in the wagger ui (e.g. D27 on the
dev deployment of phabricator has 2 diffs)

Reviewers: zalun

Reviewed By: zalun

Differential Revision: https://phabricator.services.mozilla.com/D401