auth0: add decorator to check access_token scopes.
Summary:
This change introduces the `require_auth0` class based decorator
which is able to check provided access_tokens for a set of required
scopes. The old `require_access_token` and `require_auth0_userinfo`
decorators have been refactored into this class based decorator to
keep things consistent and make it harder to screw up and leave an
endpoint insecure.

Any use of `require_auth0` will require and verify the access_token.
Additionally, the `scopes` and `userinfo` parameters to allow
specifying scope requirements and if userinfo is required and should
be fetched.

`scopes` is a required kwarg which will cause an assertion failure
if not provided. This is to make it very clear at the call site what
type of scope security is present and make the case of no scope
checking explicit and obvious (`require_auth0(scopes=())`).

`userinfo` is an optional kwarg which defaults to `False`, meaning
the userinfo will not be required or fetched automatically. To
require userinfo pass `userinfo=True`.

This change intentionally leaves out any scope enforcement on real
endpoints as this will be handled in a followup.

Test Plan: New test coverage added and `invoke test` passes.

Reviewers: imadueme

Reviewed By: imadueme

Differential Revision: https://phabricator.services.mozilla.com/D403