landings: require scopes when triggering a landing.
Summary:
The landings endpoint now requires the lando, profile, and email
scopes. In order to support this in development and testing the
MockAuth0 now automatically generates tokens with the required scopes.
To be more production like it also adds the openid scope, which all
tokens will have.

Test Plan: `invoke test` passes

Reviewers: imadueme

Reviewed By: imadueme

Differential Revision: https://phabricator.services.mozilla.com/D405