lando: Get parents with one request.
Summary:
Currently we're requesting the list of parent revisions by iterating
over the PHID. Each of the parent is received in its own request.
Instead we should request all parent revisions in one go.

Nit fixed:
Description in `get_revision` was suggesting to use D{int} as an id.
Modified it to be clear we want an integer.

Reviewers: imadueme, mars

Reviewed By: imadueme, mars

Bug #: 1433943

Differential Revision: https://phabricator.services.mozilla.com/D471