auth0: add caching for auth0 responses.
Summary:
Auth0 rate limits requests to the userinfo and jwks endpoints. In order
to avoid hitting these limits we now cache the responses for both
endpoints.

If a key is removed from the jwks endpoint we should no longer trust
tokens signed by that key. By caching this endpoint's response we now
introduce a window where we will still trust tokens signed by a revoked
key. This means our choice of cache expiry length must be a tradeoff
between performance and security. 60 seconds has been chosen, which
seems reasonable. Shortening this time in the future would be fine
as the auth0 rate limit for jwks is 20 requests per second.

The userinfo responses are also cached for 60 seconds and the cache
key is dependent on the specific token itself, rather than the user
the token represents. This distinction is important, as the custom
claims present in the response are the state of the user when the
token was generated; Changes to the underlying claims data will only
appear in the response when a token generated after the change is used.

This change includes fixes for some of the issues brought up in the
audit of lando api's auth code, since the code was already being
touched.

Test Plan: `invoke test` passes

Reviewers: dkl

Reviewed By: dkl

Subscribers: dkl, imadueme

Differential Revision: https://phabricator.services.mozilla.com/D546