app: log secret keys with a single star
Summary:
Our logs were revealing the length of our secret keys, which is a
security risk. In addition, the code was crashing on keys with a
None value since it didn't check that prior to calling len().

This commit changes the code to log with a fixed amount of stars
fixing both issues.

Reviewers: smacleod

Reviewed By: smacleod

Subscribers: smacleod, ckolos, mcote, imadueme

Bug #: 1439985

Differential Revision: https://phabricator.services.mozilla.com/D633