mocks: add __init__.py to make mocks properly importable
Summary:
When the landoapi application is installed for production it lacks
the mocks package since we don't have an __init__.py inside the folder.
This caused crashes when trying to do things like run the migration script.

Test Plan: try to run the migration script after installing the package.

Reviewers: ckolos_admin

Reviewed By: ckolos_admin

Bug #: 1437154

Differential Revision: https://phabricator.services.mozilla.com/D634