landings: add open parent & already queued blockers (bug 1440360).
Summary:
This change moves the open parent revisions and landing already queued
checks from the Landing creation to the LandingAssessment. Theses
conditions will not be checked by the dryrun endpoint.

Additionally a `@lazy` decorator has been added to allow passing "lazy
evaluated" values down into the LandingAssessment checks. This decorator
returns a callable which can be passed arguments that instead of being
applied immediately will be used to curry a new callable that will take
no arguments. This curried callable will only execute the lazy decorated
function the first time it is called, all following invocations will
return the previously calculated value.

This is particulary useful for checks that require expensive to
calculate data such as the revision data requested from phabricator.
Because fetching the revision data is now lazy it is only requested
if/when it is needed. This same lazy function is then re-used when the
actual landing code requires data about the revision.

The primary benefit of this is realized in the actual landings endpoint
(rather than the dryrun) where any blocker will cause the checks to
short curcuit and immediately return. Expensive calculations in the
later checks won't be run if an earlier check fails.

Test Plan: `invoke test` passes, new tests added

Reviewers: imadueme

Reviewed By: imadueme

Bug #: 1440360

Differential Revision: https://phabricator.services.mozilla.com/D644