landings: add warning for diff not being latest (bug 1441031).
Summary:
This change removes the force_override_diff_id special case for
acknowledging the diff to be landed isn't the latest diff. To replace it
a new LandingProblem warning for checking this case has been added.
This new check should have the same protections as the old check since
the returned message from the LandingProblem contains the diff id to be
landed as well as the diff id being overridden.

Similarly the check for the diff id not being related to the revision at
all has been moved to a LandingProblem. Similarly to the DoesNotExist
landing problem this problem raises a ProblemException immediately
return an error response to the api consumer. Moving this check here
allowed for more code to be torn out of Landing and allows for other
blockers / warnings coming after it to be simpler and assume this has
been checked.

The test coverage for overriding a specific diff id has been dropped as
this is now handled like any other warning that is acknowledged. The
case where the warning is not acknowledged is tested. In the future as
the acknowledgement story is expanded upon the test coverage for this
will be beefed up.

Test Plan: `invoke test` passes, new tests added.

Reviewers: imadueme

Reviewed By: imadueme

Bug #: 1441031

Differential Revision: https://phabricator.services.mozilla.com/D645