transplant_client: refactor and add http basic auth header
Summary:
Refactors the transplant client to better send the request, capture/log
any errors that happen, and send back to the user a generic error message.

This commit also adds an HTTP Basic Auth header that transplant requires as
well as two new environment variables to support this: TRANSPLANT_USERNAME,
and TRANSPLANT_PASSWORD.

Also included is an update to the docker-compose.yml file to add some missing
redis configuration settings (commented out) so that we can view all required
environment variables in one place.

Lastly our Sentry instance was being initialized in a way that would work when
deployed, but, not when testing (we also had no instance of sentry to reference
when we wanted to manually capture an exception). This has been fixed.

Test Plan: invoke test, feel free to test sending an actual patch to a local transplant.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1439606

Differential Revision: https://phabricator.services.mozilla.com/D647