phabricator_client: simplify get_repo methods into one
Summary:
The phabricator client had 3 methods to get a repo's information.
These have been simplified into one method: get_repo.
This method uses the /diffusion.repository.search Phabricator API method
which provides the full repo information.

This one new method now immediately returns None if no phid is given
(following the pattern of other methods in this client that return None
instead of require a value is set). This prevents the bug in 1441600
where when no phid is given phabricator performs a general search and
returns a list of all repos, then when we get the first index we
incorrectly associated that repo with a revision that had none.

Also, instead of providing separate 'full_name', 'short_name', and 'callsign'
names for the same repo, we now provide one: 'name'. This name matches
the 'name' field of the repo on Phabricator which almost always matches
the tree name on our production instance.
In addition the repo url is removed to discourage linking to the phabricaor
repo, clients should link to the repo's official location (e.g. hg.m.o).

Test Plan: run invoke test.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1441600

Differential Revision: https://phabricator.services.mozilla.com/D653