landings: map revision repo to treestatus tree
Summary:
Before submitting a landing request to the transplant client, we
must map the phabricator repository to a treestatus tree defined
at https://treestatus.mozilla-releng.net/trees. Only a subset
of the repos on phabricator will be supported initially.

Adds mapping for two repos -
https://hg.mozilla.org/automation/phabricator-qa-stage/
https://hg.mozilla.org/automation/phabricator-qa-dev/

The transplant service doesn't accept these repo's as of this commit
but will once a bug is filed to add them.

Test Plan: invoke test

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1440151

Differential Revision: https://phabricator.services.mozilla.com/D654