landings: add warning for previous landings (Bug 1380283).
Summary:
This change adds a new PreviouslyLanded LandingProblem as a warning.
This will warn if the revision attempting to be landed has been landed
successfully at any time in the past.

A warning will still be generated even if another landing request is
made after the successful landing (The alternative being only warning
if the latest landing request was successful). This will generally come
up in the case of a change landing, being backed out, and then failing
to land a second time due to a rebase error or some other issue. Since
acknowledging the warning will be pretty simple it's fine that we'll be
warning in this rare-ish case. In the future we can be smarter and
detect the backout case, dealing with it accordingly.

Test Plan: `invoke test` passes, new tests added

Reviewers: imadueme

Reviewed By: imadueme

Bug #: 1380283

Differential Revision: https://phabricator.services.mozilla.com/D659