config: require TRANSPLANT_API_KEY when PINGBACK_ENABLED (Bug 1442247).
Summary:
When PINGBACK_ENABLED="y" requests to the pingback are enabled and
require the 'API-Key' header match TRANSPLANT_API_KEY in order to
authenticate. Before this change we would set the configuration to use
a `None` value if the TRANSPLANT_API_KEY env var is not set. This is
sort of a footgun since if environment in production was messed up a
little we might all of a sudden just let requests to the pingback
endpoint go through without auth.

This change will cause the app to completely abort on startup if
pingback is enabled but we don't have a non empty api key for transplant
to use.

I also considered adding a check for the api key having some sort of
minimum length or complexity, but decided that isn't really necessary
for our deployment scenario. If we change our mind we can always add it
as a followup.

Test Plan:
`invoke test` passes and verified app works with proper
env but fails when it should.

Reviewers: imadueme

Reviewed By: imadueme

Bug #: 1442247

Differential Revision: https://phabricator.services.mozilla.com/D673