landings: add support for push_bookmark repos (Bug 1445043).
Summary:
This change refactors the hard-coded repo configuration a bit to include
support for repositories that require a 'push_bookmark' be sent to
autoland. While a repo requiring this change isn't configured currently,
we'll be able to quickly add it when Lando is rolled out to
version-control-tools.

Test Plan: `invoke test` passes with new test added.

Reviewers: glob

Reviewed By: glob

Bug #: 1445043

Differential Revision: https://phabricator.services.mozilla.com/D731