database: add command to create database, nuke migrations (Bug 1442016).
Summary:
In preperation for the first production deployment, and having real data
stored on dev, this change nukes all the current migrations and adds a
commmand for creating the db from scratch. This will make some quick
fixups to the database schema easier pre-production.

As part of adding the new command for creating a database this change
switches away from the unmaintained flask_script to the recommended and
built-in flask click cli. 'lando-cli' can be invoked to run the cli
which includes the default flask commands as well as the added 'init'
command which will create the db.

As discussed with ops when this change is deployed to dev the current
database will be removed and a new schema created with 'lando-cli init'.
This won't really be a loss of anything important since we haven't
actually used dev for any landings. The same init process will be used
on the first deploy to production.

Some migrationless DB schema tweaks will follow this change before these
deploys happen, giving us a nice fresh start.

Test Plan:
`invoke test` passes, tested created new schemas and
inspected the database after.

Reviewers: glob

Reviewed By: glob

Bug #: 1442016

Differential Revision: https://phabricator.services.mozilla.com/D773