pingback: properly handle large pingbacks and landed status.
Summary:
The Landing model was using undersized columns for a number of fields
causing failures when transplant would send things like rebase failures.
This change changes error and result columns to be Text and increases
the size of requester_email to the maximum length for a valid email.
This fixes Bug 1444841.

The way 'landed' in pingback messages is handled is also changed to be
in line with mozreview / what transplant intends. Now when transplant
sends an update message in 'result' with landed == False we treat it as
a simple update, rather than a failure. This frequently happens when
transplant tries to land to a closed tree and sends an update to inform
the user it will try again later. This fixes Bug 1444842.

Because most of the places Landings are created / saved were being
updated I took the opportunity to remove the Landing.save() method. The
code making use of models now follows the suggested session management
convention where commits are managed externally. This should make it
much easier to clean up things like the Landing race condition (Bug
1441269) later. This is the final change needed to fix Bug 1442023.

Bugs Fixed: 1444841, 1444842, 1442023

Depends on D774

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Bug #: 1444841

Differential Revision: https://phabricator.services.mozilla.com/D775