commits: import commit parsing from vct.
Summary:
version-control-tools contains useful python code for parsing commit
messages and manipulating them. Particularly we're interested in the
reviewer replacement and bug number parsing. This change copies that
code mostly verbatim from commit f0c7808d6c00[1].

I chose to copy the code as most of vct is python 2 and depending on
clone of vct for this functionality would be overkill. In the future it
would be nice to have a regularly updated py 2/3 library on pypi for
this sort of thing.

[1] https://mzl.la/2DQvLnd

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Differential Revision: https://phabricator.services.mozilla.com/D810