commit_message: properly rewrite messages (Bug 1449245).
Summary:
We now prepend "Bug <#> - " to the first line of the commit message if
we have a bug number from Phabricator that is not found in the first
line already. This means we will respect the convention of the commit
but also ensure the bug # is always present.

Reviewers are now also rewritten using the code from vct. We only
include the reviewers that have accepted the revision, rather than all
users marked as a reviewer, which was the previous behaviour. We do not
special case reviewers which have accepted an older version of the diff,
which is left to a future change.

Depends on D816.

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Bug #: 1449245

Differential Revision: https://phabricator.services.mozilla.com/D817