phabricator: respond nicely to errors (Bug 1451491).
Summary:
This change introduces a new custom error handler for
`PhabricatorAPIException`. If these exceptions are allowed to bubble up
they will result in a proper response.

As part of this change a lot of the code accessing data on responses
from phabricator has been updated to cause a `PhabricatorAPIException`
instead of things like `IndexError`. To make this and future work around
our phabricator requests easier many of the methods on the
`PhabricatorClient` have been removed.

Depends on D848.

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Bug #: 1451491

Differential Revision: https://phabricator.services.mozilla.com/D849