repos: make repo configuration per environment (Bug 1452636).
Summary:
We don't really want the same repositories enabled in dev/stage/prod so
now they are configured per environment. We only have two environments
configured for now, 'devsvcdev' and 'devscvprod', which correspond to
the value of 'ENV' in the dev and prod cloud services environments.

The 'phabricator-qa-dev' repository has been removed from Lando's
configuration as it is not configured in dev Transplant.

As part of this change repository and repository configuration existence
checks have been made part of the LandingAssessment. This should improve
the UX around visiting revision's with unsuported repositories in Lando.

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Bug #: 1452636

Differential Revision: https://phabricator.services.mozilla.com/D888