revisions: stop sending parent_revisions (Bug 1436570).
Summary:
We were requesting and sending a lot of data around the parent's of a
requested revision, which wasn't being used by Lando UI. Generating this
data can be expensive and makes maintaining tests and new features more
dificult when it is not needed.

In the future when parents are actually needed (for features that aren't
on our roadmap yet), it'll probably be best to send just the revision
IDs. We can allow a /revisions/ request which searches for specific IDs,
which will be less expensive for cases where Lando API clients don't
care about parent data.

This also changes the behaviour when checking for open parent revisions
from looking over the full dependency tree to only the direct parents.
This was discussed with the team and should be fine since closed
revisions should never really have open parents themselves. If this
becomes a real pain point we can revisit it later.

Depends on D942

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Bug #: 1436570

Differential Revision: https://phabricator.services.mozilla.com/D945