tests: revamp phabricator mocking strategy (Bug 1436570).
Summary:
In order to prepare for use of more powerful filtering / searching
features of Phabricator the PhabResponseFactory has been replaced by a
PhabricatorDouble. This new mock acts as a proper replacement for
phabricator, allowing interdependency between phabricator objects and
proper searching / filtering with requests. This change was important to
allow replacing some of the conduit calls in upcoming changes.

PhabricatorDouble can act as both a test double, generating Phabricator
data without use of the PhabricatorClient, and a monkeypatcher, which
replaces responses the PhabriccatorClient generates with mock data. This
makes it useful in both unit and integration tests.

The data generation / storage mechanism is separate from the response
generation for various conduit methods. This separation allows for
various methods that rely on the same underlying Phabricator objects to
return consistent results. This was the main motivating factor for
replacing PhabResponseFactory, as extending it to handle this would have
been non-trivial.

Some of the tests have been removed as part of this change, as they were
redundant or not actually testing what they indicated they were. It
should now be much easier to refactor the landoapi code to be more
testable with unit tests, replacing many of the integration tests as we
go.

This change should be mostly low risk as only testing code was modified.

Depends on D945

Test Plan: `invoke test` passes, tests updated / added.

Reviewers: glob

Reviewed By: glob

Bug #: 1436570

Differential Revision: https://phabricator.services.mozilla.com/D960