phabricator: use 'differential.diff.search' (Bug 1436570).
Summary:
This change goes most of the way of switching from
'differential.querydiffs' to 'differential.diff.search'. Taking care of
the rest of querydiffs usage should be a quick followup when the
'commits' attachment for 'differential.diff.search' is deployed. It made
sense to start making this change now as calculating if a review is on
an older diff requires the diff phid, which is not provided by
querydiffs.

The PhabricatorDouble has been updated to support
'differential.diff.search'. Basic 'refs'key support has been added and
can be expanded as more ref information is required.

The 'vcs_base_revision' key has been removed from the revision response
as Lando UI is not using it, and it was calculated using querydiffs.
'differential.diff.search' has richer information surrounding the
commits that are part of the diff, and that it is based on. If commit
information ends up being needed it can be added back.

Depends on D998.

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Bug #: 1436570

Differential Revision: https://phabricator.services.mozilla.com/D1008